+++ CHANGELOG.md
@@ -17,3 +17,4 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - Double sums of powers decomposition
 - Triple sums of powers decomposition
 - Multifold sums of powers
+- Multifold sums of powers mathematica